Refactor moving collection package into container
