+++ fixture-monkey-api/src/main/java/com/navercorp/fixturemonkey/api/container/LruCache.java
@@ -1,13 +1,19 @@
-package com.navercorp.fixturemonkey.api.collection;
+package com.navercorp.fixturemonkey.api.container;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
 
+import org.apiguardian.api.API;
+import org.apiguardian.api.API.Status;
+
 /**
- * It is deprecated.
- * Use {@link com.navercorp.fixturemonkey.api.container.LruCache} instead.
+ * It is the Least Recently Used cache.
+ * It would remove the least recently used element when it is full.
+ *
+ * @param <K> key of the cache
+ * @param <V> value of the cache
  */
-@Deprecated
+@API(since = "0.6.0", status = Status.EXPERIMENTAL)
 public final class LruCache<K, V> extends LinkedHashMap<K, V> {
 	private final int maxSize;
